requirements.txt
appdirs==1.4.4
asgiref==3.3.1
cachetools==4.2.0
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.3.1
cryptography==3.3.1
defusedxml==0.6.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.5
django-3-jet==1.0.8
django-allauth==0.42.0
django-cors-headers==3.6.0
django-email-verification==0.0.7
django-filter==2.4.0
django-jet==1.0.8
django-rest-auth==0.9.5
django-rest-passwordreset==1.1.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
et-xmlfile==1.0.1
excel2json==1.0.1
excel2json-3==0.1.6
filelock==3.0.12
google-api-core==1.24.1
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
inflection==0.5.1
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
Markdown==3.3.3
MarkupSafe==1.1.1
numpy==1.19.0
oauthlib==3.1.0
opencv-python==4.2.0.34
openpyxl==3.0.5
packaging==20.8
pandas==1.2.0
Pillow==7.1.2
protobuf==3.13.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.3
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
simplejwt==2.0.1
six==1.15.0
sqlparse==0.3.1
typing==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.10
virtualenv==20.0.21
whitenoise==5.1.0
wheel==0.36.2
xlrd==2.0.1